Breaking circular deps with indirect injected values
