+++ packages/realm/src/Collection.ts
@@ -17,7 +17,8 @@
 ////////////////////////////////////////////////////////////////////////////
 
 import type { binding } from "../binding";
-import type { Dictionary, DictionaryAccessor } from "./Dictionary";
+import { injectIndirect } from "./indirect";
+import type { Dictionary } from "./Dictionary";
 import type { List } from "./List";
 import type { OrderedCollectionAccessor } from "./OrderedCollection";
 import type { RealmSet } from "./Set";
@@ -26,6 +27,7 @@ import type { TypeHelpers } from "./TypeHelpers";
 import { type CallbackAdder, Listeners } from "./Listeners";
 import { IllegalConstructorError, type TypeAssertionError } from "./errors";
 import { assert } from "./assert";
+import type { DictionaryAccessor } from "./collection-accessors/Dictionary";
 
 /**
  * Collection accessor identifier.
@@ -201,3 +203,5 @@ export abstract class Collection<
 
 /* eslint-disable-next-line @typescript-eslint/no-explicit-any -- We define these once to avoid using "any" through the code */
 export type AnyCollection = Collection<any, any, any, any, any>;
+
+injectIndirect("Collection", Collection);
