+++ packages/realm/src/Dictionary.ts
@@ -18,8 +18,7 @@
 
 import { assert } from "./assert";
 import { binding } from "../binding";
-import { List, createListAccessor, insertIntoListOfMixed, isJsOrRealmList } from "./List";
-import { Results, createResultsAccessor } from "./Results";
+import { indirect, injectIndirect } from "./indirect";
 import { COLLECTION_ACCESSOR as ACCESSOR, Collection, COLLECTION_TYPE_HELPERS as TYPE_HELPERS } from "./Collection";
 import { AssertionError, IllegalConstructorError } from "./errors";
 import type { DefaultObject } from "./schema";
@@ -28,6 +27,8 @@ import type { Realm } from "./Realm";
 import { toItemType } from "./TypeHelpers";
 import type { TypeHelpers } from "./TypeHelpers";
 import { RealmObject } from "./Object";
+import type { DictionaryAccessor } from "./collection-accessors/Dictionary";
+import { createResultsAccessor } from "./collection-accessors/Results";
 
 /* eslint-disable jsdoc/multiline-blocks -- We need this to have @ts-expect-error located correctly in the .d.ts bundle */
 
@@ -215,7 +216,7 @@ export class Dictionary<T = unknown> extends Collection<
     const itemType = toItemType(values.type);
     const typeHelpers = this[TYPE_HELPERS];
     const accessor = createResultsAccessor({ realm, typeHelpers, itemType });
-    const results = new Results<T>(realm, values, accessor, typeHelpers);
+    const results = new indirect.Results<T>(realm, values, accessor, typeHelpers);
 
     for (const value of results.values()) {
       yield value;
@@ -237,7 +238,7 @@ export class Dictionary<T = unknown> extends Collection<
     const itemType = toItemType(snapshot.type);
     const typeHelpers = this[TYPE_HELPERS];
     const accessor = createResultsAccessor({ realm, typeHelpers, itemType });
-    const results = new Results<T>(realm, snapshot, accessor, typeHelpers);
+    const results = new indirect.Results<T>(realm, snapshot, accessor, typeHelpers);
 
     for (let i = 0; i < size; i++) {
       const key = keys.getAny(i);
@@ -324,126 +325,7 @@ export class Dictionary<T = unknown> extends Collection<
   }
 }
 
-/**
- * Accessor for getting and setting items in the binding collection.
- * @internal
- */
-export type DictionaryAccessor<T = unknown> = {
-  get: (dictionary: binding.Dictionary, key: string) => T;
-  set: (dictionary: binding.Dictionary, key: string, value: T) => void;
-};
-
-type DictionaryAccessorFactoryOptions<T> = {
-  realm: Realm;
-  typeHelpers: TypeHelpers<T>;
-  itemType: binding.PropertyType;
-  isEmbedded?: boolean;
-};
-
-/** @internal */
-export function createDictionaryAccessor<T>(options: DictionaryAccessorFactoryOptions<T>): DictionaryAccessor<T> {
-  return options.itemType === binding.PropertyType.Mixed
-    ? createDictionaryAccessorForMixed<T>(options)
-    : createDictionaryAccessorForKnownType<T>(options);
-}
-
-function createDictionaryAccessorForMixed<T>({
-  realm,
-  typeHelpers,
-}: Pick<DictionaryAccessorFactoryOptions<T>, "realm" | "typeHelpers">): DictionaryAccessor<T> {
-  const { toBinding, fromBinding } = typeHelpers;
-  return {
-    get(dictionary, key) {
-      const value = dictionary.tryGetAny(key);
-      switch (value) {
-        case binding.ListSentinel: {
-          const accessor = createListAccessor<T>({ realm, itemType: binding.PropertyType.Mixed, typeHelpers });
-          return new List<T>(realm, dictionary.getList(key), accessor, typeHelpers) as T;
-        }
-        case binding.DictionarySentinel: {
-          const accessor = createDictionaryAccessor<T>({ realm, itemType: binding.PropertyType.Mixed, typeHelpers });
-          return new Dictionary<T>(realm, dictionary.getDictionary(key), accessor, typeHelpers) as T;
-        }
-        default:
-          return fromBinding(value) as T;
-      }
-    },
-    set(dictionary, key, value) {
-      assert.inTransaction(realm);
-
-      if (isJsOrRealmList(value)) {
-        dictionary.insertCollection(key, binding.CollectionType.List);
-        insertIntoListOfMixed(value, dictionary.getList(key), toBinding);
-      } else if (isJsOrRealmDictionary(value)) {
-        dictionary.insertCollection(key, binding.CollectionType.Dictionary);
-        insertIntoDictionaryOfMixed(value, dictionary.getDictionary(key), toBinding);
-      } else {
-        dictionary.insertAny(key, toBinding(value));
-      }
-    },
-  };
-}
-
-function createDictionaryAccessorForKnownType<T>({
-  realm,
-  typeHelpers,
-  isEmbedded,
-}: Omit<DictionaryAccessorFactoryOptions<T>, "itemType">): DictionaryAccessor<T> {
-  const { fromBinding, toBinding } = typeHelpers;
-  return {
-    get(dictionary, key) {
-      return fromBinding(dictionary.tryGetAny(key));
-    },
-    set(dictionary, key, value) {
-      assert.inTransaction(realm);
-
-      if (isEmbedded) {
-        toBinding(value, { createObj: () => [dictionary.insertEmbedded(key), true] });
-      } else {
-        dictionary.insertAny(key, toBinding(value));
-      }
-    },
-  };
-}
-
-/** @internal */
-export function insertIntoDictionaryOfMixed(
-  dictionary: Dictionary | Record<string, unknown>,
-  internal: binding.Dictionary,
-  toBinding: TypeHelpers["toBinding"],
-) {
-  // TODO: Solve the "removeAll()" case for self-assignment (https://github.com/realm/realm-core/issues/7422).
-  internal.removeAll();
-
-  for (const key in dictionary) {
-    const value = dictionary[key];
-    if (isJsOrRealmList(value)) {
-      internal.insertCollection(key, binding.CollectionType.List);
-      insertIntoListOfMixed(value, internal.getList(key), toBinding);
-    } else if (isJsOrRealmDictionary(value)) {
-      internal.insertCollection(key, binding.CollectionType.Dictionary);
-      insertIntoDictionaryOfMixed(value, internal.getDictionary(key), toBinding);
-    } else {
-      internal.insertAny(key, toBinding(value));
-    }
-  }
-}
-
-/** @internal */
-export function isJsOrRealmDictionary(value: unknown): value is Dictionary | Record<string, unknown> {
-  return isPOJO(value) || value instanceof Dictionary;
-}
-
-/** @internal */
-export function isPOJO(value: unknown): value is Record<string, unknown> {
-  return (
-    typeof value === "object" &&
-    value !== null &&
-    // Lastly check for the absence of a prototype as POJOs
-    // can still be created using `Object.create(null)`.
-    (value.constructor === Object || !Object.getPrototypeOf(value))
-  );
-}
-
 /* eslint-disable-next-line @typescript-eslint/no-explicit-any -- We define these once to avoid using "any" through the code */
 export type AnyDictionary = Dictionary<any>;
+
+injectIndirect("Dictionary", Dictionary);
