+++ packages/realm/src/platform/node/index.ts
@@ -20,10 +20,10 @@ import "./binding";
 import "./fs";
 import "./device-info";
 import "./sync-proxy-config";
-import "./custom-inspect";
 import "./garbage-collection";
 
 import { Realm } from "../../Realm";
 export = Realm;
 
-import "./deprecated-global";
+import "./custom-inspect";
+import "../../deprecated-global";
