+++ packages/realm/src/property-accessors/Dictionary.ts
@@ -19,9 +19,10 @@
 import { binding } from "../../binding";
 import { assert } from "../assert";
 import { TypeAssertionError } from "../errors";
-import { Dictionary, createDictionaryAccessor } from "../Dictionary";
+import { Dictionary } from "../Dictionary";
 import { getTypeHelpers, toItemType } from "../TypeHelpers";
 import type { PropertyAccessor, PropertyOptions } from "./types";
+import { createDictionaryAccessor } from "../collection-accessors/Dictionary";
 
 /** @internal */
 export function createDictionaryPropertyAccessor({
