+++ packages/realm/src/property-accessors/types.ts
@@ -18,7 +18,7 @@
 
 import type { binding } from "../../binding";
 import type { ClassHelpers } from "../ClassHelpers";
-import type { ListAccessor } from "../List";
+import type { ListAccessor } from "../collection-accessors/List";
 import type { Realm } from "../Realm";
 import type { PresentationPropertyTypeName } from "../schema";
 import type { TypeHelpers } from "../TypeHelpers";
