+++ packages/realm/src/type-helpers/Mixed.ts
@@ -18,12 +18,9 @@
 
 import { binding } from "../../binding";
 import { assert } from "../assert";
-import { Collection } from "../Collection";
-import { Counter } from "../Counter";
-import { Dictionary, createDictionaryAccessor } from "../Dictionary";
-import { List, createListAccessor } from "../List";
-import { REALM, RealmObject } from "../Object";
-import { RealmSet } from "../Set";
+import { indirect } from "../indirect";
+import { createDictionaryAccessor } from "../collection-accessors/Dictionary";
+import { createListAccessor } from "../collection-accessors/List";
 import {
   boxToBindingGeospatial,
   circleToBindingGeospatial,
@@ -32,8 +29,9 @@ import {
   isGeoPolygon,
   polygonToBindingGeospatial,
 } from "../GeoSpatial";
+import { Counter } from "../Counter";
 import { getTypeHelpers } from "../TypeHelpers";
-import { OBJECT_INTERNAL } from "../symbols";
+import { OBJECT_INTERNAL, OBJECT_REALM } from "../symbols";
 import { TYPED_ARRAY_CONSTRUCTORS } from "./array-buffer";
 import type { TypeHelpers, TypeOptions } from "./types";
 
@@ -63,14 +61,14 @@ export function mixedToBinding(
     return null;
   } else if (value instanceof Date) {
     return binding.Timestamp.fromDate(value);
-  } else if (value instanceof RealmObject) {
+  } else if (value instanceof indirect.Object) {
     if (value.objectSchema().embedded) {
       throw new Error(`Using an embedded object (${value.constructor.name}) as ${displayedType} is not supported.`);
     }
-    const otherRealm = value[REALM].internal;
+    const otherRealm = value[OBJECT_REALM].internal;
     assert.isSameRealm(realm, otherRealm, "Realm object is from another Realm");
     return value[OBJECT_INTERNAL];
-  } else if (value instanceof RealmSet || value instanceof Set) {
+  } else if (value instanceof indirect.Set || value instanceof Set) {
     throw new Error(`Using a ${value.constructor.name} as ${displayedType} is not supported.`);
   } else if (value instanceof Counter) {
     let errMessage = `Using a Counter as ${displayedType} is not supported.`;
@@ -78,7 +76,7 @@ export function mixedToBinding(
     throw new Error(errMessage);
   } else {
     if (isQueryArg) {
-      if (value instanceof Collection || Array.isArray(value)) {
+      if (value instanceof indirect.Collection || Array.isArray(value)) {
         throw new Error(`Using a ${value.constructor.name} as a query argument is not supported.`);
       }
       // Geospatial types can currently only be used when querying and
@@ -123,11 +121,16 @@ function mixedFromBinding(options: TypeOptions, value: binding.MixedArg): unknow
   } else if (value instanceof binding.List) {
     const mixedType = binding.PropertyType.Mixed;
     const typeHelpers = getTypeHelpers(mixedType, options);
-    return new List(realm, value, createListAccessor({ realm, typeHelpers, itemType: mixedType }), typeHelpers);
+    return new indirect.List(
+      realm,
+      value,
+      createListAccessor({ realm, typeHelpers, itemType: mixedType }),
+      typeHelpers,
+    );
   } else if (value instanceof binding.Dictionary) {
     const mixedType = binding.PropertyType.Mixed;
     const typeHelpers = getTypeHelpers(mixedType, options);
-    return new Dictionary(
+    return new indirect.Dictionary(
       realm,
       value,
       createDictionaryAccessor({ realm, typeHelpers, itemType: mixedType }),
